requirements.txt
streamlit
streamlit-option-menu
streamlit-aggrid
xlsxwriter
fitz
accelerate
PyMuPDF
google-generativeai
